rework arch structure
