+++ frontend/src/lib/components/SearchBar.svelte
