+++ frontend/src/lib/models/Audio.ts
@@ -1,0 +1,5 @@
+export interface AudioQuality {
+	maximumBitDepth: number;
+	maximumSamplingRate: number;
+	isHiRes: boolean;
+}
