+++ frontend/src/lib/models/Image.ts
@@ -1,0 +1,6 @@
+export interface Image {
+	small: string;
+	thumbnail: string;
+	large: string;
+	back?: string | null;
+}
