+++ frontend/src/lib/models/Response.ts
@@ -1,0 +1,5 @@
+import type { Track } from '$lib/models/Track';
+
+export interface SearchResponse {
+	tracks: Track[];
+}
