+++ frontend/src/lib/models/Track.ts
@@ -1,0 +1,33 @@
+import type { AudioQuality } from '$lib/models/Audio';
+import type { Image } from '$lib/models/Image';
+
+export interface Track {
+	id: number;
+	title: string;
+	artist: string;
+	artistId: number;
+	albumTitle: string;
+	albumCover: string;
+	albumId: string;
+	releaseDate: string;
+	genre: string;
+	duration: number;
+	audioQuality: AudioQuality;
+	version: string | null;
+	label: string;
+	labelId: number;
+	upc: string;
+	mediaCount: number;
+	parental_warning: boolean;
+	streamable: boolean;
+	purchasable: boolean;
+	previewable: boolean;
+	genreId: number;
+	genreSlug: string;
+	genreColor: string;
+	releaseDateStream: string;
+	releaseDateDownload: string;
+	maximumChannelCount: number;
+	images: Image;
+	isrc: string;
+}
